FAKECAM: Capture only requests exceptions
As written, the python script captures all exceptions
and ignores them. This makes things pretty annoying when
we want to use CTRL+C or other system calls to kill the
process as that loop captures those exceptions as well.

This patch addresses the issue by limiting the try/catch
to only catch requests.RequestExceptions.

Signed-off-by: Mark Stenglein <mark@stengle.in>